% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data.R
\docType{data}
\name{aal116coordinates}
\alias{aal116coordinates}
\title{AAL116 brain atlas coordinates in MNI space}
\format{116 observations (Brain Region Names) of 7 variables (name, x.mni, y.mni, z.mni, lobe, hemi, index)}
\usage{
data(aal116coordinates)
}
\description{
Automated Anatomical Labeling (AAL): The AAL atlas distributed with the AAL Toolbox was fractionated to functional
resolution (3x3x3 mm3) using nearest-neighbor interpolation.
This data is available at \url{http://preprocessed-connectomes-project.org/abide/Pipelines.html} as part of ABIDE-preprocessed dataset.
It can be directly downloaded at \url{https://fcp-indi.s3.amazonaws.com/data/Projects/ABIDE_Initiative/Resources/aal_roi_atlas.nii.gz}
}
\references{
Cameron Craddock, Yassine Benhajali, Carlton Chu, Francois Chouinard, Alan Evans, András Jakab, Budhachandra Singh Khundrakpam, John David Lewis, Qingyang Li, Michael Milham, Chaogan Yan, Pierre Bellec (2013). The Neuro Bureau Preprocessing Initiative: open sharing of preprocessed neuroimaging data and derivatives. In Neuroinformatics 2013, Stockholm, Sweden.
}
\keyword{datasets}
