% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data.R
\docType{data}
\name{ABIDE_aal116_timeseries}
\alias{ABIDE_aal116_timeseries}
\title{ABIDE I preprocessed time series grouped by control and autism and partitioned by AAL116 atlas}
\format{a list of two data matrices of time series(1:2250, 1:116) and (1:2060,1:116)}
\usage{
data(ABIDE_aal116_timeseries)
}
\description{
This time series data is available as part of Autism Brain Imaging Data Exchange (ABIDE).
ABIDE is a collaboration of 16 international imaging sites that have aggregated and are openly
sharing neuroimaging data from 539 individuals suffering from ASD and 573 typical controls.
For data access, please refer to \url{http://preprocessed-connectomes-project.org/abide/download.html}.
The data is preprocessed, cancatnated and organized into two data matrices for easy input.
}
\references{
Cameron Craddock, Yassine Benhajali, Carlton Chu, Francois Chouinard, Alan Evans, András Jakab, Budhachandra Singh Khundrakpam, John David Lewis, Qingyang Li, Michael Milham, Chaogan Yan, Pierre Bellec (2013). The Neuro Bureau Preprocessing Initiative: open sharing of preprocessed neuroimaging data and derivatives. In Neuroinformatics 2013, Stockholm, Sweden.
}
\keyword{datasets}
