% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_name_to_out.R
\name{add_name_to_out}
\alias{add_name_to_out}
\title{helper function to add row/col names to JointNets precision matrix output
To help label igraph object in returngraph and plot}
\usage{
add_name_to_out(output, datalist, ...)
}
\arguments{
\item{output}{output of jointnets}

\item{datalist}{orginial data list}

\item{...}{unused}
}
\value{
output with names from datalist
}
\description{
helper function to add row/col names to JointNets precision matrix output
To help label igraph object in returngraph and plot
}
