% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AUC.R
\name{AUC}
\alias{AUC}
\title{function to return AUC score}
\usage{
AUC(simulate, method, lambdas, ...)
}
\arguments{
\item{simulate}{results from function simulation}

\item{method}{method name from any one of the jointNets methods}

\item{...}{parameters passed to method}

\item{range}{values of lambda to calculate AUC score}
}
\value{
a list consisting of AUC (a list of AUC scores), precision (a list of precisions)
and recall (a list of recalls)
}
\description{
function to return AUC score
}
