% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluation.R
\name{BIC}
\alias{BIC}
\title{function to return BIC score}
\usage{
BIC(datalist, result)
}
\arguments{
\item{datalist}{datalist used as an input to any of the jointnet method}

\item{result}{result generated from datalist using the same jointnet method}
}
\value{
BIC score (lower scores correspond to better estimation)
}
\description{
function to return BIC score
}
