% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BIC.R
\name{BIC_matrix}
\alias{BIC_matrix}
\title{function to calculate BIC score of a list of estimated precision matrices}
\usage{
BIC_matrix(covs, graphs, n, p)
}
\arguments{
\item{covs}{-- a list of empirical covariance matrices}

\item{graphs}{-- a list of estimated precision matrices}

\item{n}{-- total number of samples (total number of samples from all lists)}

\item{p}{-- number of features}
}
\value{
BIC score
}
\description{
function to calculate BIC score of a list of estimated precision matrices
}
