% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comparisonplot.R
\name{comparisonplot.diffee}
\alias{comparisonplot.diffee}
\title{function to plot comparsion plot between result and simulationresult graph}
\usage{
\method{comparisonplot}{diffee}(result, simulationresult)
}
\description{
function to plot comparsion plot between result and simulationresult graph
}
\examples{
\dontrun{
library(JointNets)
simulationresult = simulation(p = 20,  n = c(100,100))
result = diffee(simulationresult$simulatedsamples[[1]], simulationresult$simulatedsamples[[2]], 1)
comparisonplot(result,simulationresult)
}
}
