% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comparisonplot.R
\name{comparisonplot.diffeek}
\alias{comparisonplot.diffeek}
\title{function to plot comparsion plot between result and simulationresult graph}
\usage{
\method{comparisonplot}{diffeek}(result, simulationresult)
}
\description{
function to plot comparsion plot between result and simulationresult graph
}
