% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comparisonplot.R
\name{comparisonplot.fasjem}
\alias{comparisonplot.fasjem}
\title{function to plot comparsion plot between result and simulationresultd graph}
\usage{
\method{comparisonplot}{fasjem}(result, simulationresult)
}
\description{
function to plot comparsion plot between result and simulationresultd graph
}
\examples{
\dontrun{
library(JointNets)
simulationresult = simulation(p = 20,  n = c(100,100))
result = fasjem(simulationresult$simulatedsamples, method = "fasjem-g", 2, 0.1, 0.1, 0.05, 20)
comparisonplot(result,simulationresult)
}
}
