% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comparisonplot.R
\name{comparisonplot.jeek}
\alias{comparisonplot.jeek}
\title{function to plot comparsion plot between result and simulationresult graph}
\usage{
\method{comparisonplot}{jeek}(result, simulationresult)
}
\description{
function to plot comparsion plot between result and simulationresult graph
}
\examples{
\dontrun{
library(JointNets)
simulationresult = simulation(p = 20,  n = c(100,100))
result = jeek(simulationresult$simulatedsamples,0.25,covType = "kendall",parallel = TRUE)
comparisonplot(result,simulationresult)
}
}
