% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluation.R
\name{comparisonplot}
\alias{comparisonplot}
\title{function to plot comparsion plot between result and simulated graph}
\usage{
comparisonplot(result, simulate)
}
\arguments{
\item{result}{result generated from any of the jointnet methods}

\item{simulate}{simulated result from simulateGraph}
}
\description{
function to plot comparsion plot between result and simulated graph
}
