% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comparisonplot.R
\name{comparisonplot.wsimule}
\alias{comparisonplot.wsimule}
\title{function to plot comparsion plot between result and simulationresultd graph}
\usage{
\method{comparisonplot}{wsimule}(result, simulationresult)
}
\description{
function to plot comparsion plot between result and simulationresultd graph
}
\examples{
\dontrun{
library(JointNets)
simulationresult = simulation(p = 20,  n = c(100,100))
result = wsimule(simulationresult$simulatedsamples, 0.2, 1, W = matrix(1,20,20), covType = "cov", TRUE)
comparisonplot(result,simulationresult)
}
}
