% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_cov.R
\name{compute_cov}
\alias{compute_cov}
\title{helper function to add compute covariance matrix / kendall tau correlation matrix}
\usage{
compute_cov(X, covType = "cov")
}
\arguments{
\item{X}{data matrix}

\item{covType}{"cov" or "kendall"}
}
\value{
covriance matrix / kendall tau correlation matrix
}
\description{
helper function to add compute covariance matrix / kendall tau correlation matrix
}
