% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nii_input.R
\name{convert_nii_to_datalist}
\alias{convert_nii_to_datalist}
\title{convert nii data to datalist according to an atlas mask}
\usage{
convert_nii_to_datalist(data, atlas)
}
\arguments{
\item{data}{data from fmri scan (4d array)}

\item{atlas}{atlas of corresponding size (3d array)}
}
\description{
convert nii data to datalist according to an atlas mask
}
\details{
need to test extensively on this
}
