% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diffeek.R
\name{diffeek}
\alias{diffeek}
\title{Fast and Scalable Learning of Sparse Changes in High-Dimensional Gaussian
Graphical Model Structure}
\usage{
diffeek(C, D, W, g, epsilon = 1, lambda = 0.05, covType = "cov",
  thre = "soft")
}
\arguments{
\item{C}{A input matrix for the 'control' group. It can be data matrix or
covariance matrix. If C is a symmetric matrix, the matrices are assumed to
be covariance matrix. More details at \url{https://github.com/QData/DIFFEE}}

\item{D}{A input matrix for the 'disease' group. It can be data matrix or
covariance matrix. If D is a symmetric matrix, the matrices are assumed to
be covariance matrix. More details at \url{https://github.com/QData/DIFFEE}}

\item{g}{grouping information}

\item{epsilon}{A positive number}

\item{lambda}{A positive number. The hyperparameter controls the sparsity
level of the matrices. The \eqn{\lambda_n} in the following section:
Details.}

\item{covType}{A parameter to decide which Graphical model we choose to
estimate from the input data.

If covType = "cov", it means that we estimate multiple sparse Gaussian
Graphical models. This option assumes that we calculate (when input X
represents data directly) or use (when X elements are symmetric representing
covariance matrices) the sample covariance matrices as input to the simule
algorithm.

If covType = "kendall", it means that we estimate multiple nonparanormal
Graphical models. This option assumes that we calculate (when input X
represents data directly) or use (when X elements are symmetric representing
correlation matrices) the kendall's tau correlation matrices as input to the
simule algorithm.}

\item{thre}{A parameter to decide which threshold function to use for
\eqn{T_v}. If thre = "soft", it means that we choose soft-threshold function
as \eqn{T_v}. If thre = "hard", it means that we choose hard-threshold
function as \eqn{T_v}.}

\item{w}{weight matrix}
}
\value{
\item{diffNet}{A matrix of the estimated sparse changes between two
Gaussian Graphical Models}
}
\description{
The DIFFEEK algorithm
}
\references{
Beilun Wang, Arshdeep Sekhon, Yanjun Qi (2018). Fast and
Scalable Learning of Sparse Changes in High-Dimensional Gaussian Graphical
Model Structure. \url{arXiv:1710.11223}
}
\author{
Beilun Wang
}
