% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimension_reduce.R
\name{dimension_reduce}
\alias{dimension_reduce}
\title{reduce the dimensionality of the datalist if needed}
\usage{
dimension_reduce(datalist)
}
\arguments{
\item{datalist}{a datalist of high dimensionality}
}
\value{
a datalist of reduced dimensionality
}
\description{
reduce the dimensionality of the datalist if needed
}
\examples{
library(JointNets)
data(exampleData)
reduction = dimension_reduce(exampleData)
}
