% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluation.R
\name{evaluation}
\alias{evaluation}
\title{function to evaluate results from jointnet (from simulated data)}
\usage{
evaluation(result, simulate)
}
\arguments{
\item{result}{result generated using the same jointnet method}

\item{simulate}{result generated from function simulate}
}
\description{
function to evaluate results from jointnet (from simulated data)
}
