% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/F1.R
\name{F1.fasjem}
\alias{F1.fasjem}
\title{computes F1 score for jointnet result}
\usage{
\method{F1}{fasjem}(result, simulatedgraphs, ...)
}
\description{
computes F1 score for jointnet result
}
\examples{
\dontrun{
simulationresult = simulation(p = 20,  n = c(100,100))
truth = simulationresult$simulatedgraphs
result = fasjem(simulationresult$simulatedsamples, method = "fasjem-g", 2, 0.1, 0.1, 0.05, 20)
F1(result,truth)
}
}
