% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/F1.R
\name{F1.jeek}
\alias{F1.jeek}
\title{computes F1 score for jointnet result}
\usage{
\method{F1}{jeek}(result, simulatedgraphs, ...)
}
\description{
computes F1 score for jointnet result
}
\examples{
\dontrun{
simulationresult = simulation(p = 20,  n = c(100,100))
truth = simulationresult$simulatedgraphs
result = jeek(simulationresult$simulatedsamples,0.25,covType = "kendall",parallel = TRUE)
F1(result,truth)
}
}
