% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluation.R
\name{F1}
\alias{F1}
\title{function to calculate F1 score}
\usage{
F1(simulate, result)
}
\arguments{
\item{simulate}{result from simulateGraph}

\item{result}{result from any of the jointnets method using the samples generated by simulate}
}
\value{
a list of F1 scores between simulate and result between 0 and 1 (higher scores correspond to better estimation)
}
\description{
function to calculate F1 score
}
\details{
scores' order correspond to task order and the last score in the list represent score for shared graph (only applicable to simule and wsimule)
}
