% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/F1.R
\name{F1.simule}
\alias{F1.simule}
\title{computes F1 score for jointnet result}
\usage{
\method{F1}{simule}(result, simulatedgraphs, ...)
}
\description{
computes F1 score for jointnet result
}
\examples{
\dontrun{
simulationresult = simulation(p = 20,  n = c(100,100))
truth = simulationresult$simulatedgraphs
result = simule(simulationresult$simulatedsamples, 0.2, 0.5, covType = "cov", TRUE)
F1(result,truth)
}
}
