% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/F1.R
\name{F1_single}
\alias{F1_single}
\title{function to calculate F1 score for a single simulate graph with a single result graph}
\usage{
F1_single(truth, estimate)
}
\arguments{
\item{truth}{ground truth precision matrix}

\item{estimate}{estimated precision matrix}
}
\value{
a single F1 score between truth and estimate between 0 and 1 (higher scores correspond to better estimation)
}
\description{
function to calculate F1 score for a single simulate graph with a single result graph
}
