% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{generateSampleList}
\alias{generateSampleList}
\title{function to generate a list of samples from simulatedGraph result}
\usage{
generateSampleList(simulate, n)
}
\arguments{
\item{simulate}{result from simulateGraph}

\item{n}{a vector of corresponding size to indicate number of samples for each task}
}
\value{
a list of length(n) data matrices
}
\description{
function to generate a list of samples from simulatedGraph result
}
\details{
if n is c(100,200,300) and p is 20, the function will return a list of 3 data matrices of size (100x20,200x20,300x20)
}
