% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{generateSamples}
\alias{generateSamples}
\title{function to generate samples from a single precision matrix}
\usage{
generateSamples(precision, n = 100)
}
\arguments{
\item{precision}{pxp precision matrix (generated from simulateGraph)}

\item{n}{number of samples}
}
\value{
a list of nXp randomly generated gaussian samples from pxp precision matrix
}
\description{
function to generate samples from a single precision matrix
}
