% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intertwined.R
\name{intertwined}
\alias{intertwined}
\title{calculate intertwined covariance matrix}
\usage{
intertwined(X, covType = "cov")
}
\arguments{
\item{X}{a list of covariance matrix}
}
\value{
intertwined covariance matrices
}
\description{
calculate intertwined covariance matrix
}
