% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{jointplot}
\alias{jointplot}
\title{core function for plotting jeek, simule, fasjem and wsimule
need to fix diffee title issue}
\usage{
jointplot(x, type = "task", neighbouroption = "task", subID = NULL,
  index = NULL, hastitle = TRUE, haslegend = TRUE, ...)
}
\description{
core function for plotting jeek, simule, fasjem and wsimule
need to fix diffee title issue
}
