% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{make.adj.matrix}
\alias{make.adj.matrix}
\title{helper function to make adjacency matrix}
\usage{
make.adj.matrix(theta, separate = FALSE)
}
\arguments{
\item{theta}{result from jointnets}
}
\description{
helper function to make adjacency matrix
}
