% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_adj_matrix.R
\name{make_adj_matrix}
\alias{make_adj_matrix}
\title{method to make adjancey matrix for outputs from different methods}
\usage{
make_adj_matrix(x, separate = FALSE)
}
\description{
method to make adjancey matrix for outputs from different methods
}
