% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{maketitle}
\alias{maketitle}
\title{helper function to make title}
\usage{
maketitle(x, type = "task", subID = NULL, index = NULL,
  neighbouroption = "task", graphlabel = NULL)
}
\description{
helper function to make title
}
