% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotbrain.R
\name{plotbrain}
\alias{plotbrain}
\title{plot 3d brain network from JointNets result}
\usage{
plotbrain(x, ...)
}
\arguments{
\item{x}{output generated from any one of the JointNets functions (simule,wsimule,jeek,fasjem,diffee,diffeek)}

\item{...}{additional arguments, please see \code{\link{plotbrain.simule}}, \code{\link{plotbrain.wsimule}} and etc for details}
}
\value{
3d (rgl) brain network
}
\description{
This function plots 3d brain network from JointNets result
}
\details{
The function plots brain network using \code{\link{rglplot.igraph}}
}
\examples{
\dontrun{
library(JointNets)
graphics.off()
par(ask=F)
par(mfrow=c(1,1))
data(ABIDE_aal116_timeseries)
data(aal116coordinates)
layout = cbind(aal116coordinates$x.mni + 90,
aal116coordinates$y.mni+126, aal116coordinates$z.mni+72)
result = simule(ABIDE_aal116_timeseries, 0.2, 1, covType = "cov", TRUE)
plotbrain(result, type = "task", neighbouroption = "task",
subID = NULL, index = NULL, layout = layout)
}
}
\author{
Zhaoyang Wang (Author), Zhaoyang Wang (maintainer) \email{zw4dn@virginia.edu}
}
