% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotbrain.R
\name{plotbrain.wsimule}
\alias{plotbrain.wsimule}
\title{plot 3d brain network from wsimule result}
\usage{
\method{plotbrain}{wsimule}(x, type = "task", neighbouroption = "task",
  subID = NULL, index = NULL, hastitle = TRUE, haslegend = TRUE,
  ...)
}
\arguments{
\item{x}{output generated from wsimule}

\item{...}{additional arguments, please see \code{\link{plot.wsimule}} for details}

\item{layout}{numeric matrix (3 columns) that specifies node positions in 3d brain network (atlas coordinates)}
}
\value{
3d (rgl) brain network
}
\description{
This function plots 3d brain network from wsimule result
}
\details{
The function plots brain network using \code{\link{rglplot.igraph}}
}
\examples{
\dontrun{
library(JointNets)
graphics.off()
par(ask=F)
par(mfrow=c(1,1))
data(ABIDE_aal116_timeseries)
data(aal116coordinates)
layout = cbind(aal116coordinates$x.mni + 90,aal116coordinates$y.mni+126, aal116coordinates$z.mni+72)
result = wsimule(ABIDE_aal116_timeseries, 0.2, 1, W = matrix(1,116,116), covType = "cov", TRUE)
plotbrain(result, type = "task", neighbouroption = "task", subID = NULL, index = NULL, layout = layout)
}
}
\author{
Zhaoyang Wang (Author), Zhaoyang Wang (maintainer) \email{zw4dn@virginia.edu}
}
