% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/returngraph.R
\name{return_graph_K_joint}
\alias{return_graph_K_joint}
\title{function to return gadj and K (number of tasks)}
\usage{
return_graph_K_joint(x)
}
\description{
function to return gadj and K (number of tasks)
}
