% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/returngraph.R
\name{returngraph.simulation}
\alias{returngraph.simulation}
\title{function returngraph for simulated result}
\usage{
\method{returngraph}{simulation}(x, type = "task",
  neighbouroption = "task", subID = NULL, index = NULL)
}
\description{
function returngraph for simulated result
}
