% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/returngraph.R
\name{returngraph_jointnets}
\alias{returngraph_jointnets}
\title{core function to return graph from jointnets result}
\usage{
returngraph_jointnets(x, type = "task", neighbouroption = "task",
  subID = NULL, index = NULL)
}
\description{
core function to return graph from jointnets result
}
