% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{simulateGraph}
\alias{simulateGraph}
\title{function to simulate multiple sparse graphs}
\usage{
simulateGraph(p = 20, N = 2, seedNum = 37, s = 0.1, ss = 0.1)
}
\arguments{
\item{p}{number of features}

\item{N}{number of tasks}

\item{seedNum}{seed number for random simulation}

\item{s}{controls sparsity of the generated graph}

\item{ss}{controls sparsity of the generated graph}
}
\value{
a list of N related sparse pXp precision matrices (graphs)
}
\description{
function to simulate multiple sparse graphs
}
