% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{simulation}
\alias{simulation}
\title{function to simulate multiple sparse graphs and generate samples}
\usage{
simulation(p = 20, n, seedNum = 37, s = 0.1, ss = 0.1)
}
\arguments{
\item{p}{number of features}

\item{n}{a vector indicating number of samples and tasks, for example c(100,200,300) for 3 tasks and 100,200 and 300 samples per task}

\item{seedNum}{seed number for random simulation}

\item{s}{controls sparsity of the generated graph}

\item{ss}{controls sparsity of the generated graph}
}
\value{
a list comprising graphs (multiple related simulated graphs) and samples (samples generated from multiple related graphs)
}
\description{
function to simulate multiple sparse graphs and generate samples
}
