% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AUC.R
\name{AUC}
\alias{AUC}
\title{function to return AUC score for JointNets methods given}
\usage{
AUC(simulationresult, gm_method = "simule", lambdas, ...)
}
\arguments{
\item{simulationresult}{results from function simulation}

\item{gm_method}{method name from any one of the jointNets methods}

\item{lambdas}{a vector of lambda values for the JointNets method to run with}

\item{...}{extra parameters passed to the JointNets method}
}
\value{
AUC score, a list of precisions and recalls
}
\description{
function to return AUC score for JointNets methods given
}
\examples{
\dontrun{
simulationresult = simulation(n=c(100,100,100))
AUC_result = AUC(simulationresult,lambdas = seq(0.1,2,0.05),epsilon = 2)
AUC_result
plot(AUC_result$fPM,AUC_result$tPM)
}
}
