% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BIC.R
\name{BIC}
\alias{BIC}
\title{function to return BIC score}
\usage{
BIC(datalist, result)
}
\arguments{
\item{datalist}{datalist used as an input to any of the jointnet method}

\item{result}{result generated from datalist using the same jointnet method}
}
\value{
BIC score
}
\description{
function to return BIC score
}
\details{
not working with DIFFEE and DIFFEEK (difference estimation)
}
\examples{
\dontrun{
simulateresult = simulation(p = 20,  n = c(100,100))
result = simule(simulateresult$simulatedsamples, 0.2, 0.5, covType = "cov", TRUE)
BIC(simulateresult$simulatedsamples,result)
}
}
