% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BIC.R
\name{BIC}
\alias{BIC}
\title{calculate BIC score for JointNets method}
\usage{
BIC(datalist, result)
}
\arguments{
\item{datalist}{datalist used as an input to any of the JointNets method}

\item{result}{result generated from datalist using the same JointNets method}
}
\value{
BIC score
}
\description{
calculate BIC score for JointNets method
}
\details{
not working with DIFFEE and kdiffnet (difference estimation)
}
\examples{
library(JointNets)
simulateresult = simulation(p = 20,  n = c(100,100))
result = simule(simulateresult$simulatedsamples, 0.2, 0.5, covType = "cov", FALSE)
BIC(simulateresult$simulatedsamples,result)
}
\author{
Zhaoyang Wang (Author), Zhaoyang Wang (maintainer) \email{zw4dn@virginia.edu}
}
