% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comparisonplot.R
\name{comparisonplot}
\alias{comparisonplot}
\title{function to plot comparsion plot between result and simulationresultd graph}
\usage{
comparisonplot(result, simulationresult)
}
\arguments{
\item{result}{result generated from any of the jointnet methods}

\item{simulationresult}{simulationresultd result from simulationresultGraph}
}
\description{
function to plot comparsion plot between result and simulationresultd graph
}
\examples{
\dontrun{
library(JointNets)
simulationresult = simulation(p = 20,  n = c(100,100))
result = simule(simulationresult$simulatedsamples, 0.2, 0.5, covType = "cov", TRUE)
comparisonplot(result,simulationresult)
}
}
