% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/F1.R
\name{F1.diffee}
\alias{F1.diffee}
\title{computes F1 score for jointnet result}
\usage{
\method{F1}{diffee}(result, simulatedgraphs, ...)
}
\arguments{
\item{result}{output generated from any one of the jointnet algorithms}

\item{simulatedgraphs}{$simulatedgraphs from function simulation()}

\item{...}{unused}
}
\description{
computes F1 score for jointnet result
}
\examples{
library(JointNets)
simulationresult = simulation(p = 20,  n = c(100,100))
truth = simulationresult$simulatedgraphs
result = diffee(simulationresult$simulatedsamples[[1]],
simulationresult$simulatedsamples[[2]], 0.01)
F1(result,truth)
}
