% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/F1.R
\name{F1}
\alias{F1}
\title{This function computes F1 score for jointnet result}
\usage{
F1(result, simulatedgraphs, ...)
}
\arguments{
\item{result}{output generated from any one of the jointnet algorithms}

\item{...}{additional arguments}
}
\value{
computed F1 scores
}
\description{
This function computes F1 score for jointnet result
}
\examples{
\dontrun{
}
}
\author{
Zhaoyang Wang (Author), Zhaoyang Wang (maintainer) \email{zw4dn@virginia.edu}
}
