% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluation.R
\name{F1_single}
\alias{F1_single}
\title{function to calculate F1 score for a single simulate graph with a single result graph}
\usage{
F1_single(simulate, result)
}
\arguments{
\item{simulate}{single precision matrix of simulateGraph}

\item{result}{single corresponding precision matrix generated by any of the jointnets method}
}
\value{
a single F1 score between simulate and result between 0 and 1 (higher scores correspond to better estimation)
}
\description{
function to calculate F1 score for a single simulate graph with a single result graph
}
