% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_adj_matrix.R
\name{make_adj_matrix}
\alias{make_adj_matrix}
\title{method to make adjancey matrix for outputs from different methods}
\usage{
make_adj_matrix(x, separate = FALSE)
}
\arguments{
\item{x}{result from JointNet methods}

\item{separate}{determines whther the adj matrix is accumulative or individual}
}
\value{
a single adj matrix or a list of adj matrices
}
\description{
method to make adjancey matrix for outputs from different methods
}
