% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_title.R
\name{maketitle}
\alias{maketitle}
\title{make title for different graph estimation}
\usage{
maketitle(x, ...)
}
\value{
a descriptive title for the plotted graph
}
\description{
make title for different graph estimation
}
