% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_title.R
\name{maketitle}
\alias{maketitle}
\title{S3 method to make title for different graph estimation}
\usage{
maketitle(x, ...)
}
\value{
according to the type of graph, return a descriptive title
}
\description{
S3 method to make title for different graph estimation
}
