% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotbrain.R
\name{plotbrain.diffeek}
\alias{plotbrain.diffeek}
\title{plot 3d brain network from diffeek result}
\usage{
\method{plotbrain}{diffeek}(x, ...)
}
\arguments{
\item{x}{output generated from diffee}

\item{...}{additional arguments}
}
\value{
3d (rgl) brain network
}
\description{
This function plots 3d brain network from diffee result
}
\details{
The function plots brain network using \code{\link{rglplot.igraph}}
}
\examples{
\dontrun{
}
}
\author{
Zhaoyang Wang (Author), Zhaoyang Wang (maintainer) \email{zw4dn@virginia.edu}
}
